# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: DYL Real-Estate Services API
  description: >
    Admin-verified real estate listings API (Lagos-first, Nigeria). Three
    roles share this API: **seekers** (browse/save/inquire, the default
    `role` on every account), **owners** (list/manage their own
    properties — a seeker becomes an owner the moment they create a
    listing), and **admins** (`role: admin`) who review and verify
    listings and moderate the platform.

    Authentication is a bearer JWT returned by `/auth/login`,
    `/auth/register`, or `/auth/google`. Send it as
    `Authorization: Bearer <token>` on every authenticated request.
    Tokens are invalidated (not just expired) when the account's
    password changes — see `token_version` under Auth below.
  version: "1.0.0"
servers:
  - url: /api
    description: Relative to wherever this API is deployed (see README for
      environment-specific base URLs).
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Obtained from `/auth/login`, `/auth/register`, or `/auth/google`.
        Carries `id`, `role`, `email`, `name`, and `token_version` — the
        last of which is checked against the user's current value on
        every request, so it's what lets a password change or reset
        immediately invalidate every previously-issued token rather than
        waiting out the token's natural expiry.
  parameters:
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
      description: 1-indexed page number.
...
  responses:
    BadRequest:
      description: Validation failed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing, invalid, expired, or revoked token.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string, example: "Invalid email or password." }
    Pagination:
      type: object
      properties:
        page: { type: integer, example: 1 }
        limit: { type: integer, example: 20 }
        total: { type: integer, example: 47 }
        totalPages: { type: integer, example: 3 }
...
paths:
...
  /properties:
    get:
      tags: [Properties]
      summary: Search public listings
      description: Always excludes `status != active` and
        `verification_status = rejected` — a rejected listing is never
        publicly visible regardless of other filters. Page size caps at 50.
      security: []
      parameters:
        - name: q
          in: query
          schema: { type: string }
          description: Free-text search across title, description, city, state.
        - name: state
          in: query
          schema: { type: string }
        - name: city
          in: query
          schema: { type: string }
        - name: property_type
          in: query
          schema: { type: string, enum: [land, residential, commercial] }
        - name: listing_type
          in: query
          schema: { type: string, enum: [sale, rent] }
        - name: min_price
          in: query
          schema: { type: number }
        - name: max_price
          in: query
          schema: { type: number }
        - name: verified_only
          in: query
          schema: { type: boolean }
        - name: has_video
          in: query
          schema: { type: boolean }
        - $ref: "#/components/parameters/PageParam"
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 12 }
      responses:
        "200":
          description: OK — possibly an empty `data` array; never an error
            for "no matches".
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Property" } }
                  pagination: { $ref: "#/components/schemas/Pagination" }
    post:
      tags: [Properties]
      summary: Create a listing
      description: Owned by the authenticated caller; always starts
        `verification_status = pending` regardless of what's sent.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PropertyCreate" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                type: object
                properties: { data: { $ref: "#/components/schemas/Property" } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /properties/market/movers:
    get:
      tags: [Properties]
      summary: Recent price gainers/losers
      description: Computed from recorded `price_history` — only listings
        with an actual documented price change appear here.
      security: []
      parameters:
        - name: days
          in: query
          schema: { type: integer, minimum: 1, maximum: 365, default: 90 }
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema: { type: object }
  /properties/me/listings:
    get:
      tags: [Properties]
      summary: The current user's own listings
      description: Includes the caller's own pending/rejected listings
        (unlike the public search, which hides those). Max page size 100.
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 100 }
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Property" } }
                  pagination: { $ref: "#/components/schemas/Pagination" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /properties/me/saved:
    get:
      tags: [Properties]
      summary: The current user's saved listings
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 100 }
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Property" } }
                  pagination: { $ref: "#/components/schemas/Pagination" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /properties/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
    get:
      tags: [Properties]
      summary: Get one listing
      description: A pending listing is visible to anyone; a rejected one
        is only visible to its owner or an admin (404 for everyone else,
        including "does this ID exist" — no distinction is made between
        "doesn't exist" and "exists but you can't see it").
      security: []
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                type: object
                properties: { data: { $ref: "#/components/schemas/Property" } }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Properties]
      summary: Update a listing
      description: Owner or admin only. When a **non-admin** edits a
        `verified` listing, `verification_status` is automatically reset
        to `pending` — any edit re-queues it for review. An admin's edits
        never trigger that reset.
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PropertyCreate" }
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema:
                type: object
                properties: { data: { $ref: "#/components/schemas/Property" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: Authenticated, but not this listing's owner (and not an admin).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Properties]
      summary: Delete a listing
      description: Owner or admin only.
      responses:
        "204": { description: Deleted. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: Authenticated, but not this listing's owner (and not an admin).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404": { $ref: "#/components/responses/NotFound" }
  /properties/{id}/save:
    post:
      tags: [Properties]
      summary: Toggle save/unsave
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Toggled.
          content:
            application/json:
              schema:
                type: object
                properties: { saved: { type: boolean } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /properties/{id}/inquiries:
    post:
      tags: [Properties]
      summary: Send an inquiry about a listing
      description: Public (login optional — if authenticated, the inquiry
        is linked to the account). Rate-limited. Emails the listing's
        owner (best-effort; a delivery failure doesn't fail the request).
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, message]
              properties:
                name: { type: string }
                email: { type: string, format: email }
                phone: { type: string }
                message: { type: string }
      responses:
        "201":
          description: Sent.
          content:
            application/json:
              schema:
                type: object
                properties: { message: { type: string } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
...
  /admin/properties:
    get:
      tags: [Admin]
      summary: List all listings (any status)
      description: Unlike public search, includes pending/rejected/sold —
        this is the review queue. Max page size 100.
      parameters:
        - name: verification_status
          in: query
          schema: { type: string, enum: [pending, verified, rejected] }
        - $ref: "#/components/parameters/PageParam"
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 100 }
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Property" } }
                  pagination: { $ref: "#/components/schemas/Pagination" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...